+++ src/components/datasets-list/datasets-item-header/index.tsx
@@ -1,16 +1,21 @@
 'use client';
+
+import { FC } from 'react';
+
 import Info from '@/components/datasets-list/datasets-item-info';
 import Icon from '@/components/icon/component';
 import DOWNLOAD_SVG from 'svgs/ui/download.svg?sprite';
 import SHARE_SVG from 'svgs/ui/share.svg?sprite';
-export const DatasetsItemHeader = ({
-  layer_id,
-  title,
-  downloadUrlBase,
-}: {
+
+type DatasetsItemHeaderTypes = {
   layer_id: string;
   title: string;
   downloadUrlBase?: string;
+};
+export const DatasetsItemHeader: FC<DatasetsItemHeaderTypes> = ({
+  layer_id,
+  title,
+  downloadUrlBase,
 }) => {
   return (
     <div className="flex items-start justify-between">