+++ src/components/datasets-list/datasets-item/index.tsx
@@ -1,30 +1,44 @@
 'use client';
-import { useCallback, useMemo } from 'react';
+import { FC, useCallback, useMemo } from 'react';
 
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
 import { cn } from '@/lib/classnames';
+
+import type { LayerSettingTypes } from '@//types/layers';
 
 import { useLayerSource } from '@/hooks/map';
 
 import ItemHeader from '@/components/datasets-list/datasets-item-header';
 import { Button } from '@/components/ui/button';
 
-export const DatasetsItem = ({ layer_id }: { layer_id: string }) => {
+export const DatasetsItem: FC<{ layer_id: string }> = ({ layer_id }) => {
   const { data } = useLayerSource({ layer_id });
   const { title, description, download_url } = data;
   const router = useRouter();
   const pathname = usePathname();
   const params = useSearchParams();
   const layersId = params.get('layers');
-  const layerParams = useMemo<string[]>(
-    () => layersId?.substring(1, layersId?.length - 1).split(','),
-    [layersId]
-  ) satisfies string[];
-  const isActive = layerParams?.includes(layer_id);
+  const layersIdParsed = useMemo<null | LayerSettingTypes[]>(() => {
+    if (layersId === null) return null;
+    else return JSON.parse(layersId) as LayerSettingTypes[];
+  }, [layersId]);
+  const layerId = useMemo<LayerSettingTypes['id']>(() => layersIdParsed?.[0]?.id, [layersIdParsed]);
+
+  const isActive = layerId === layer_id;
+
+  // Create the layers object
+  const layersObject = { id: layer_id, opacity: 1 };
+
+  // Encode the layers object as a JSON string
+  const encodedLayers = decodeURIComponent(JSON.stringify(layersObject));
+
+  // Construct the URL
+  const url = `${pathname}/?layers=[${encodedLayers}]`;
+
   const handleClick = useCallback(() => {
-    isActive ? router.replace(`${pathname}`) : router.replace(`${pathname}/?layers=[${layer_id}]`);
-  }, [pathname, router, isActive, layer_id]);
+    isActive ? router.replace(`${pathname}`) : router.replace(url);
+  }, [pathname, router, isActive, url]);
 
   return (
     <li key={layer_id} className="space-y-6 border-b border-b-brand-200 p-7.5 ">