+++ src/components/datasets-list/index.tsx
@@ -5,7 +5,7 @@ import { LayerTypes } from '@/types/datasets';
 import DatasetsItem from './datasets-item';
 
 export const DatasetsList = ({ data }: { data: LayerTypes[] }) => (
-  <div className="pl-4 text-secondary-200">
+  <div className="pl-4 text-secondary-500">
     <ul>
       {data.map(({ layer_id }) => (
         <DatasetsItem key={layer_id} layer_id={layer_id} />