+++ src/types/layers.d.ts
@@ -1,0 +1,4 @@
+export type LayerSettingTypes = {
+  id: string;
+  opacity: number;
+};